requirements.txt
python>=3.12.6
discord.py>=2.3.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
websocket-client>=1.7.0
requests>=2.31.0